[gcc] SpellsLvl07: Make float to int conversion explicit

+++ src/game/magic/spells/SpellsLvl07.cpp
@@ -191,7 +191,7 @@ void FlyingEyeSpell::Update() {
 	eyeball.floating *= 10.f;
 	
 	if(m_lastupdate - m_timcreation <= ArxDurationMs(3000)) {
-		eyeball.exist = toMs(m_lastupdate - m_timcreation) * (1.0f / 30);
+		eyeball.exist = long(toMs(m_lastupdate - m_timcreation) * (1.0f / 30));
 		eyeball.size = Vec3f(1.f - float(eyeball.exist) * 0.01f);
 		eyeball.angle.setYaw(eyeball.angle.getYaw() + toMs(framediff3) * 0.6f);
 	} else {
